chore: Make a function class and import to main
